[finishes #161404822]refactored views to deduct quantities after sales record
